requirements.txt
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
Flask==2.0.3
idna==3.7
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==5.2.2
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
pyspark==2.4.3
python-dateutil==2.9.0.post0
python-docx==0.8.11
pytz==2024.1
requests==2.27.1
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.18
Werkzeug==2.0.3
zipp==3.6.0
